Moved quiet flag into output
