+++ cmd/analyze.go
@@ -21,6 +21,7 @@ import (
 	"fmt"
 	"os"
 
+	"github.com/GoogleCloudPlatform/container-diff/cmd/util/output"
 	"github.com/GoogleCloudPlatform/container-diff/differs"
 	pkgutil "github.com/GoogleCloudPlatform/container-diff/pkg/util"
 	"github.com/sirupsen/logrus"
@@ -38,7 +39,6 @@ var analyzeCmd = &cobra.Command{
 		return nil
 	},
 	Run: func(cmd *cobra.Command, args []string) {
-		pkgutil.SetQuiet(quiet)
 		if err := analyzeImage(args[0], types); err != nil {
 			logrus.Error(err)
 			os.Exit(1)
@@ -87,7 +87,7 @@ func analyzeImage(imageName string, analyzerArgs []string) error {
 		return fmt.Errorf("Error performing image analysis: %s", err)
 	}
 
-	pkgutil.PrintToStdErr("Retrieving analyses")
+	output.PrintToStdErr("Retrieving analyses")
 	outputResults(analyses)
 
 	if save {
@@ -100,4 +100,5 @@ func analyzeImage(imageName string, analyzerArgs []string) error {
 func init() {
 	RootCmd.AddCommand(analyzeCmd)
 	addSharedFlags(analyzeCmd)
+	output.AddFlags(analyzeCmd)
 }
