+++ cmd/diff.go
@@ -22,6 +22,7 @@ import (
 	"os"
 	"sync"
 
+	"github.com/GoogleCloudPlatform/container-diff/cmd/util/output"
 	"github.com/GoogleCloudPlatform/container-diff/differs"
 	pkgutil "github.com/GoogleCloudPlatform/container-diff/pkg/util"
 	"github.com/GoogleCloudPlatform/container-diff/util"
@@ -42,7 +43,6 @@ var diffCmd = &cobra.Command{
 		return nil
 	},
 	Run: func(cmd *cobra.Command, args []string) {
-		pkgutil.SetQuiet(quiet)
 		if err := diffImages(args[0], args[1], types); err != nil {
 			logrus.Error(err)
 			os.Exit(1)
@@ -83,7 +83,7 @@ func diffImages(image1Arg, image2Arg string, diffArgs []string) error {
 	var wg sync.WaitGroup
 	wg.Add(2)
 
-	pkgutil.PrintToStdErr("Starting diff on images %s and %s, using differs: %s\n", image1Arg, image2Arg, diffArgs)
+	output.PrintToStdErr("Starting diff on images %s and %s, using differs: %s\n", image1Arg, image2Arg, diffArgs)
 
 	imageMap := map[string]*pkgutil.Image{
 		image1Arg: {},
@@ -113,7 +113,7 @@ func diffImages(image1Arg, image2Arg string, diffArgs []string) error {
 		defer pkgutil.CleanupImage(*imageMap[image2Arg])
 	}
 
-	pkgutil.PrintToStdErr("Computing diffs")
+	output.PrintToStdErr("Computing diffs")
 	req := differs.DiffRequest{
 		Image1:    *imageMap[image1Arg],
 		Image2:    *imageMap[image2Arg],
@@ -125,7 +125,7 @@ func diffImages(image1Arg, image2Arg string, diffArgs []string) error {
 	outputResults(diffs)
 
 	if filename != "" {
-		pkgutil.PrintToStdErr("Computing filename diffs")
+		output.PrintToStdErr("Computing filename diffs")
 		err := diffFile(imageMap[image1Arg], imageMap[image2Arg])
 		if err != nil {
 			return err
@@ -152,4 +152,5 @@ func init() {
 	diffCmd.Flags().StringVarP(&filename, "filename", "f", "", "Set this flag to the path of a file in both containers to view the diff of the file. Must be used with --types=file flag.")
 	RootCmd.AddCommand(diffCmd)
 	addSharedFlags(diffCmd)
+	output.AddFlags(diffCmd)
 }
