+++ cmd/root.go
@@ -39,7 +39,6 @@ var json bool
 var save bool
 var types diffTypes
 var noCache bool
-var quiet bool
 
 var LogLevel string
 
@@ -210,7 +209,6 @@ func (d *diffTypes) Type() string {
 
 func addSharedFlags(cmd *cobra.Command) {
 	cmd.Flags().BoolVarP(&json, "json", "j", false, "JSON Output defines if the diff should be returned in a human readable format (false) or a JSON (true).")
-	cmd.Flags().BoolVarP(&quiet, "quiet", "q", false, "Suppress output to stderr.")
 	cmd.Flags().VarP(&types, "type", "t", "This flag sets the list of analyzer types to use. Set it repeatedly to use multiple analyzers.")
 	cmd.Flags().BoolVarP(&save, "save", "s", false, "Set this flag to save rather than remove the final image filesystems on exit.")
 	cmd.Flags().BoolVarP(&util.SortSize, "order", "o", false, "Set this flag to sort any file/package results by descending size. Otherwise, they will be sorted by name.")
